requirements.txt
python-dotenv>=1.0.1,<2.0.0
pydantic>=2.7.0,<3.0.0
typing-extensions>=4.10.0,<5.0.0

openai>=1.0.0,<2.0.0
streamlit>=1.36.0,<2.0.0
langchain>=0.2.0,<1.0.0
langchain-openai>=0.1.0,<1.0.0
matplotlib>=3.8.0,<4.0.0
PyYAML>=6.0.0,<7.0.0
